+++ zia/resource_zia_url_filtering_rules_test.go
@@ -194,6 +194,9 @@ resource "%s" "%s" {
     order = 1
 	url_categories = ["ANY"]
     protocols = ["ANY_RULE"]
+	device_trust_levels = [	"UNKNOWN_DEVICETRUSTLEVEL", "LOW_TRUST", "MEDIUM_TRUST", "HIGH_TRUST" ]
+	user_agent_types = [	"OPERA", "FIREFOX", "MSIE", "MSEDGE", "CHROME", "SAFARI", "MSCHREDGE", "OTHER" ]
+	user_risk_score_levels = ["LOW", "MEDIUM", "HIGH", "CRITICAL"]
     request_methods = [ "CONNECT", "DELETE", "GET", "HEAD", "OPTIONS", "OTHER", "POST", "PUT", "TRACE"]
 	location_groups {
 		id = [data.zia_location_groups.sdwan_can.id, data.zia_location_groups.sdwan_usa.id]